Added some GUI to the boards
Added a board pic to the panel background image. Took all the buttons and made them transparent and named/lined them up to the board image.  Since pictureBox control causes uses so much ram maybe use the buttons just for display purposes. Also added some basic buttons to the forms. When running the program the 'SinglePlayerForm' is set as the main form right now for testing feel free to change if needed