Fixed the buttons. Added button functionality to the WelcomeForm buttons.
Rearranged the buttons in the panel so they are in order starting from the bottom left and working up.  Added button functionality to the WelcomeForm buttons so they open up each of the other forms. They only hide the main form. So program doesn't close yet when any form is closed only the WelcomeForm