Added to CellData; GUI for Stats Form; Reading StatsData.txt; Added function to StatsForm.cs; Moved boardArray;
Added 'Button' object to the overloaded constructed to be used in the future like the example given in class. Finished GUI for StatsForm. Added function to read form text file and populate the textBoxes with data. Moved boardArray to singleplayer and twoplayer forms. Added an overloaded constructor to call pass the boardArray to the GameBoard class.