Edited DropPieces; Moved WinMessage; Moved ComputerTurn(); made WhoHasWon public;
Fixed the issue with GameOverForm not getting the correct playterTurn. Moved ComputerTurn so the AI moves after a WinChecker is ran to see if the player has won. Added Set/Get function to set and get the tempPlayerTurn so it get the player turn before the (++, --) happens. Added anoter label to the GameOverForm.